+++ changelog.md
@@ -1,5 +1,5 @@
 ## Upcoming
-* 
+* Added Absorption Armor, Armor rework and regen absorption
 
 ## 5.0.0.0-alpha
 * Ported Knockback and Critical Hits features